requirements.txt
requests
beautifulsoup4
tinydb
python-telegram-bot
